apis/resume-scorer/requirements.txt
flask
supabase
python-dotenv
fastapi
uvicorn
